backend/src/reddis/requirements.txt
astroid==2.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==4.4.2
idna==3.3
imageio==2.16.0
imageio-ffmpeg==0.4.5
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
moviepy==1.0.3
numpy==1.22.2
parso==0.8.3
pika==1.2.0
Pillow==9.0.1
platformdirs==2.5.1
proglog==0.1.9
pylint==2.12.2
pymongo==4.0.1
requests==2.27.1
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.8
wrapt==1.13.3
psycopg2-binary
redis